nes_data: use uint64_t for savestate union alignment instead of double
The mapper-state union used a `double align` member solely to force 8-byte
alignment. Switch it to uint64_t, which has the same size and alignment on all
supported targets, so the struct carries no floating-point type. Layout and
savestate binary format are unchanged.

+++ nes_emu/nes_data.h
@@ -85,7 +85,7 @@ struct mapper_state_t
 {
 	int size;
 	union {
-		double align;
+		uint64_t align; /* force 8-byte alignment (was double; no float needed) */
 		uint8_t data [max_mapper_state_size];
 	};
